requirements.txt
pandas
psycopg2-binary
sqlalchemy
python-dotenv
openpyxl
ipykernel
matplotlib
seaborn
streamlit